Non functional tweaks based on feedback.
