+++ lib/src/models/axi4_bfm/axi4_bfm.dart
@@ -1,10 +1,13 @@
 // Copyright (C) 2024-2025 Intel Corporation
 // SPDX-License-Identifier: BSD-3-Clause
 
-export 'axi4_compliance_checker.dart';
 export 'axi4_main.dart';
-export 'axi4_main_driver.dart';
-export 'axi4_monitor.dart';
+export 'axi4_main_read_driver.dart';
+export 'axi4_main_write_driver.dart';
 export 'axi4_packet.dart';
+export 'axi4_read_compliance_checker.dart';
+export 'axi4_read_monitor.dart';
 export 'axi4_subordinate.dart';
 export 'axi4_tracker.dart';
+export 'axi4_write_compliance_checker.dart';
+export 'axi4_write_monitor.dart';
